fix: Consolidate build config types
